+++ internal/assets/doc.go
@@ -1,0 +1,2 @@
+// Package assets provides numerous renderers for different formats.
+package assets
