--- internal/marked/doc.go
@@ -1,2 +1,0 @@
-// Package marked provides numerous renderers for different formats.
-package marked
